// packages/ptcg-play/src/app/table/prompt/prompt-card-item.interface.ts
import { SortableSpec } from '@ng-dnd/sortable';
import { Card } from '@ryuu-play/ptcg-server';

export const PromptCardType = 'PROMPT_CARD';

export interface PromptItem {
  card: Card;
  index: number;
  isAvailable: boolean;
  isSecret: boolean;
  scanUrl: string;
  spec?: SortableSpec<PromptItem, any>;
}